Fix French dictionary
Previously, a definitely weird (potentially broken?) dictionary-fr
was pulled from `openoffice.org`.  The source of that download is
provided by `dicollecte.org`, and I switched the dictionary to pull
that down instead.

Previously, `fr-moderne` was used as the source of the dictionary.
As dicollecte suggest using `fr-classique` instead, the dictionary
is now switched to the classique one.

Closes GH-4.